+++ examples/resources/aiven_pg_user/import.sh
@@ -1,0 +1,1 @@
+terraform import aiven_pg_user.user project/service_name/username