+++ internal/schemautil/ipfilter.go
@@ -9,6 +9,8 @@ import (
 // the same order as defined in the TF manifest.
 func NormalizeIpFilter(tfUserConfig interface{}, userConfig []map[string]interface{}) []map[string]interface{} {
 	tfInt, ok := tfUserConfig.([]interface{})
+	userConfig[0]["ip_filter"] = toStringSlice(userConfig[0]["ip_filter"].([]interface{}))
+
 	if !ok || len(tfInt) == 0 {
 		return userConfig
 	}