+++ CHANGELOG.md
@@ -11,6 +11,7 @@ nav_order: 1
 - Fix `GetServiceUserValidateFunc`
 - Fix provider panics on `terraform import` with invalid vpc peering id
 - Fix Azure vpc peering connection import
+- Add OpenSearch external integration endpoint
 - Add `aiven_pg_user` import example to docs
 
 ## [3.8.0] - 2022-09-30