+++ CHANGELOG.md
@@ -11,6 +11,7 @@ nav_order: 1
 - Fix `GetServiceUserValidateFunc`
 - Fix provider panics on `terraform import` with invalid vpc peering id
 - Fix Azure vpc peering connection import
+- Fix project create/update with `add_account_owners_admin_access` field
 
 ## [3.8.0] - 2022-09-30