+++ CHANGELOG.md
@@ -14,6 +14,7 @@ nav_order: 1
 - Fix project create/update with `add_account_owners_admin_access` field
 - Add OpenSearch external integration endpoint
 - Add `aiven_pg_user` import example to docs
+- Extend converter for the service user configuration options `ip_filter` object format
 
 ## [3.8.0] - 2022-09-30